requirements.txt
streamlit
python-dotenv
pypdf
PyMuPDF
uuid
sentence-transformers
langchain
langchain-community
langchain-text-splitters
langchain-core
langchain-groq
langchain-huggingface
langchain-pinecone
pinecone
pypdf
